pyproject.toml
[project]
name = "xecai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "pydantic>=2.12.5",
]

[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["xecai*"]

[project.optional-dependencies]
aws = ["aiobotocore>=3.2.0", "boto3>=1.42.55"]
google = ["google-genai>=1.65.0"]
anthropic = ["anthropic>=0.84.0"]
openai = ["openai>=2.24.0"]
redis = ["redis>=5.0.0"]
postgresql = ["psycopg[binary,pool]>=3.1.18", "pgvector>=0.2.5"]
pinecone = ["pinecone>=8.1.0"]

all = [
    "aiobotocore>=3.2.0",
    "boto3>=1.42.55",
    "google-genai>=1.65.0",
    "anthropic>=0.84.0",
    "openai>=2.24.0",
    "redis>=5.0.0",
    "motor>=3.3.2",
    "psycopg[binary,pool]>=3.1.18",
    "pinecone>=8.1.0",
    "pgvector>=0.2.5"
]
